- Added new configuration function     - new configurations options:         + use_hash_filename: Whether or not use hash filename         + autoload: Whether or not use custom autoloader. If you already has a autoloader, this should be set to FALSE         + check_ttl: Check if the template is new every ($check_ttl) seconds         + check_set, check_get: Functions to save flag variable (with ttl) across request (memcached, apc, xcache, etc)         + on_compile: callback which is executed when the compiler is loaded, useful to load custom filters & tags

+++ lib/Haanga.php
@@ -50,8 +50,13 @@ class Haanga
     protected static $cache_dir;
     protected static $templates_dir='.';
     protected static $debug;
-    protected static $onCompile = array();
-    protected static $onCompileTriggered = FALSE;
+    protected static $onCompile = NULL;
+    protected static $check_ttl;
+    protected static $check_get;
+    protected static $check_set;
+    protected static $use_autoload  = TRUE;
+    protected static $hash_filename = TRUE;
+
     public static $has_compiled;
 
     private function __construct()
@@ -59,48 +64,86 @@ class Haanga
         /* The class can't be instanced */
     }
 
-    final public static function RegisterAutoLoad()
-    {
-        set_include_path(get_include_path().PATH_SEPARATOR.dirname(__FILE__));
-        spl_autoload_register(array(__CLASS__, 'AutoLoad'));
-    }
-
     final public static function AutoLoad($class)
     {
         static $loaded = array();
-        static $incPaths  = array();
-        static $incPathstr;
+        static $path;
 
         if (!isset($loaded[$class]) && substr($class, 0, 6) === 'Haanga' && !class_exists($class, false)) {
-            $file           = str_replace('_', DIRECTORY_SEPARATOR, $class).'.php';
-            $incPath        = get_include_path();
+            if ($path === NULL) {
+                $path = dirname(__FILE__);
+            }
+            $file = $path.DIRECTORY_SEPARATOR.str_replace('_', DIRECTORY_SEPARATOR, $class).'.php';
+            if (is_file($file)) {
+                require $file;
+            }
             $loaded[$class] = TRUE;
-
-            if ($incPath != $incPathstr) {
-                $incPathstr = $incPath;
-                $incPaths   = explode(PATH_SEPARATOR, $incPath);
-            }
-            
-
-            foreach ($incPaths as $dir) {
-                if (file_exists($dir. DIRECTORY_SEPARATOR.$file)) {
-                    require_once $file;
-                    return TRUE;
-                }
-            }
+            return;
         }
 
         return FALSE;
     }
 
-    function onCompile($function)
+    // configure(Array $opts) {{{
+    /**
+     *  Configuration to load Haanga
+     *
+     *  Options:
+     *
+     *      - (string)   cache_dir 
+     *      - (string)   tempalte_dir
+     *      - (callback) on_compile
+     *      - (boolean)  debug
+     *      - (int)      check_ttl
+     *      - (callback) check_get
+     *      - (callback) check_set
+     *      - (boolean)  autoload
+     *      - (boolean)  use_hash_filename
+     *
+     *  @return void
+     */
+    final public static function configure(Array $opts)
     {
-        if (!is_callable($function)) {
-            throw new Haanga_Exception('Invalid callback');
+        foreach ($opts as $option => $value) {
+            switch (strtolower($option)) {
+            case 'cache_dir':
+                self::setCacheDir($value);
+                break;
+            case 'template_dir':
+                self::setTemplateDir($value);
+                break;
+            case 'on_compile':
+                if (is_callable($value)) {
+                    self::$onCompile = $value;
+                }
+                break;
+            case 'debug':
+                self::enableDebug((bool)$value);
+                break;
+            case 'check_ttl':
+                self::$check_ttl = (int)$value;
+                break;
+            case 'check_get':
+                if (is_callable($value)) {
+                    self::$check_get = $value;
+                }
+                break;
+            case 'check_set':
+                if (is_callable($value)) {
+                    self::$check_set = $value;
+                }
+                break;
+            case 'autoload':
+                self::$use_autoload = (bool)$value;
+                break;
+            case 'use_hash_filename':
+                self::$hash_filename = (bool)$value;
+            default:
+                continue;
+            }
         }
-
-        self::$onCompile[] = $function;
     }
+    // }}}
 
     // setCacheDir(string $dir) {{{
     /**
@@ -167,52 +210,69 @@ class Haanga
     public static function Load($file, $vars = array(), $return=FALSE, $blocks=array())
     {
         static $compiler;
-        if (empty(self::$cache_dir) || empty(self::$templates_dir)) {
+        if (empty(self::$cache_dir)) {
             throw new Haanga_Exception("Cache dir or template dir is missing");
         }
 
         self::$has_compiled = FALSE;
 
         $tpl      = self::$templates_dir.'/'.$file;
-        $fnc      =  sha1($tpl);
-        $php      = self::$cache_dir.'/'.$fnc.'.php';
+        $fnc      = sha1($tpl);
         $callback = "haanga_".$fnc;
+        $php = self::$cache_dir.'/'.(self::$hash_filename ? $fnc : str_replace(DIRECTORY_SEPARATOR, '_', $file)).'.php';
 
         if (is_callable($callback)) {
             return $callback($vars, $return, $blocks);
         }
 
-        if (!is_file($php) && !is_file($tpl)) {
-            throw new Exception("View {$file} doesn't exists");
-        }
+        $check = TRUE;
 
-        if (!is_file($php) || filemtime($tpl) > filemtime($php)) {
-            if (!$compiler) {
-                $compiler = new Haanga_Compiler_Runtime;
+        if (self::$check_ttl && self::$check_get && self::$check_set) {
+            /* */
+            if (call_user_func(self::$check_get, $callback)) {
+                /* disable checking for the next $check_ttl seconds */
+                $check = FALSE;
+            } else {
+                $result = call_user_func(self::$check_set, $callback, TRUE, self::$check_ttl);
             }
+        } 
+        
+        if (!is_file($php) || ($check && filemtime($tpl) > filemtime($php))) {
+
+            if (!is_file($tpl)) {
+                /* There is no template nor compiled file */
+                throw new Exception("View {$file} doesn't exists");
+            }
+
+            /* recompile */
+            if (!$compiler) {
+                /* load compiler (done just once) */
+                if (self::$use_autoload) {
+                    spl_autoload_register(array(__CLASS__, 'AutoLoad'));
+                }
+
+                $compiler = new Haanga_Compiler_Runtime;
+
+                if (self::$onCompile) {
+                    /* call onCompile hook, just the first time */
+                    call_user_func(self::$onCompile);
+                }
+            }
+
             $compiler->reset();
+
             if (self::$debug) {
                 $compiler->setDebug($php.".dump");
             }
 
-            if (count(self::$onCompile) > 0 && !self::$onCompileTriggered) {
-                foreach (self::$onCompile as $fnc) {
-                    $fnc();
-                }
-                self::$onCompileTriggered = TRUE;
-            }
-
-            //try {
-                $code = $compiler->compile_file($tpl, FALSE);
-            //} catch (Exception $e) {
-            //    throw new Haanga_Compiler_Exception($tpl.' :'.$e->getMessage().' on '.$e->getfile(). ':'.$e->Getline());
-            //}
+            $code = $compiler->compile_file($tpl, FALSE);
 
             file_put_contents($php, "<?php".$code);
             self::$has_compiled = TRUE;
         }
+
         if (!is_callable($callback)) {
-            require_once $php;
+            require $php;
         }
 
         return $callback($vars, $return, $blocks);
